adding control hiding for tabs on explore view
