+++ app/javascript/components/visualization/PlotDisplayControls.js
@@ -15,7 +15,7 @@ export const defaultExploreParams = {
 }
 
 /** the graph customization controls for the exlore tab */
-export default function RenderControls({ exploreParams, updateExploreParams }) {
+export default function RenderControls({ shownTab, exploreParams, updateExploreParams }) {
   const [showScatter, setShowScatter] = useState(false)
   const [showHeatmap, setShowHeatmap] = useState(false)
   const [showDistribution, setShowDistribution] = useState(false)
@@ -40,90 +40,68 @@ export default function RenderControls({ exploreParams, updateExploreParams }) {
   }
   return (
     <div className="render-controls">
-      <Panel className="controls-scatter" expanded={showScatter} onToggle={() => setShowScatter(!showScatter)}>
-        <Panel.Heading onClick={() => setShowScatter(!showScatter)}>
-          <Panel.Title className="action"
-            componentClass="a"
-            title="toggle scatter display controls"
-            data-analytics-name="plot-params-scatter-toggle">
-            <FontAwesomeIcon className="fa-lg" icon={showScatter ? faCaretDown : faCaretRight}/>&nbsp;
+      <Panel className={shownTab === 'scatter' ? '' : 'hidden'}>
+        <Panel.Heading>
+          <Panel.Title>
             Scatter
           </Panel.Title>
         </Panel.Heading>
-        <Panel.Collapse>
-          <Panel.Body>
-            <label htmlFor="colorscale-picker">Continuous color scale
-              <span className="detail"> (for numeric data)</span>
-            </label>
-            <Select name="colorscale-picker"
-              options={SCATTER_COLOR_OPTIONS.map(opt => ({ label: opt, value: opt }))}
-              value={{ label: scatterColorValue, value: scatterColorValue }}
-              clearable={false}
-              onChange={option => updateExploreParams({ scatterColor: option.value })}/>
-          </Panel.Body>
-        </Panel.Collapse>
+        <Panel.Body>
+          <label htmlFor="colorscale-picker">Continuous color scale
+            <span className="detail"> (for numeric data)</span>
+          </label>
+          <Select name="colorscale-picker"
+            options={SCATTER_COLOR_OPTIONS.map(opt => ({ label: opt, value: opt }))}
+            value={{ label: scatterColorValue, value: scatterColorValue }}
+            clearable={false}
+            onChange={option => updateExploreParams({ scatterColor: option.value })}/>
+        </Panel.Body>
       </Panel>
-      <Panel className="controls-distribution"
-        expanded={showDistribution}
-        onToggle={() => setShowDistribution(!showDistribution)}>
-        <Panel.Heading onClick={() => setShowDistribution(!showDistribution)}>
-          <Panel.Title className="action"
-            componentClass="a"
-            title="toggle distribution display controls"
-            data-analytics-name="plot-params-distribution-toggle">
-            <FontAwesomeIcon className="fa-lg" icon={showDistribution ? faCaretDown : faCaretRight}/>&nbsp;
+      <Panel className={shownTab === 'distribution' ? '' : 'hidden'}>
+        <Panel.Heading>
+          <Panel.Title>
             Distribution
           </Panel.Title>
         </Panel.Heading>
-        <Panel.Collapse>
-          <Panel.Body>
-            <label htmlFor="distribution-plot-picker">Plot type </label>
-            <Select name="distribution-plot-picker"
-              options={DISTRIBUTION_PLOT_OPTIONS}
-              value={distributionPlotValue}
-              clearable={false}
-              isSearchable={false}
-              onChange={option => updateExploreParams({ distributionPlot: option.value })}/>
-            <label htmlFor="distribution-plot-picker">Data points </label>
-            <Select name="distribution-points-picker"
-              options={DISTRIBUTION_POINTS_OPTIONS}
-              value={distributionPointsValue}
-              clearable={false}
-              isSearchable={false}
-              onChange={option => updateExploreParams({ distributionPoints: option.value })}/>
-          </Panel.Body>
-        </Panel.Collapse>
+        <Panel.Body>
+          <label htmlFor="distribution-plot-picker">Plot type </label>
+          <Select name="distribution-plot-picker"
+            options={DISTRIBUTION_PLOT_OPTIONS}
+            value={distributionPlotValue}
+            clearable={false}
+            isSearchable={false}
+            onChange={option => updateExploreParams({ distributionPlot: option.value })}/>
+          <label htmlFor="distribution-plot-picker">Data points </label>
+          <Select name="distribution-points-picker"
+            options={DISTRIBUTION_POINTS_OPTIONS}
+            value={distributionPointsValue}
+            clearable={false}
+            isSearchable={false}
+            onChange={option => updateExploreParams({ distributionPoints: option.value })}/>
+        </Panel.Body>
       </Panel>
-      <Panel className="controls-heatmap"
-        expanded={showHeatmap}
-        onToggle={() => setShowHeatmap(!showHeatmap)}>
-        <Panel.Heading onClick={() => setShowHeatmap(!showHeatmap)}>
-          <Panel.Title className="action"
-            componentClass="a"
-            title="toggle heatmap display controls"
-            data-analytics-name="plot-params-heatmap-toggle">
-            <FontAwesomeIcon className="fa-lg" icon={showHeatmap ? faCaretDown : faCaretRight}/>&nbsp;
+      <Panel className={shownTab === 'heatmap' ? '' : 'hidden'}>
+        <Panel.Heading>
+          <Panel.Title>
             Heatmap
           </Panel.Title>
         </Panel.Heading>
-        <Panel.Collapse>
-          <Panel.Body>
-            <label htmlFor="row-centering-picker">Row centering </label>
-            <Select name="row-centering-picker"
-              options={ROW_CENTERING_OPTIONS}
-              value={heatmapRowCenteringValue}
-              clearable={false}
-              isSearchable={false}
-              onChange={option => updateExploreParams({ heatmapRowCentering: option.value })}/>
-            <label htmlFor="fit-picker">Fit options </label>
-            <Select name="fit-picker"
-              options={FIT_OPTIONS}
-              value={heatmapFitValue}
-              clearable={false}
-              isSearchable={false}
-              onChange={option => updateExploreParams({ heatmapFit: option.value })}/>
-          </Panel.Body>
-        </Panel.Collapse>
+        <Panel.Body>
+          <label htmlFor="row-centering-picker">Row centering </label>
+          <Select name="row-centering-picker"
+            options={ROW_CENTERING_OPTIONS}
+            value={heatmapRowCenteringValue}
+            clearable={false}
+            isSearchable={false}
+            onChange={option => updateExploreParams({ heatmapRowCentering: option.value })}/>
+          <label htmlFor="fit-picker">Fit options </label>
+          <Select name="fit-picker"
+            options={FIT_OPTIONS}
+            value={heatmapFitValue}
+            clearable={false}
+            isSearchable={false}
+            onChange={option => updateExploreParams({ heatmapFit: option.value })}/>
+        </Panel.Body>
       </Panel>
     </div>
   )
