+++ app/javascript/styles/_relatedGenesIdeogram.scss
@@ -22,8 +22,6 @@
 
     #_ideogramMiddleWrap {
       border-bottom: 1px solid #EEE;
-      border-left: 1px solid #DDD;
-      border-right: 1px solid #DDD;
     }
   }
 }
